Fix Makefile to work in Emacs ansi-term
In ansi-term, the environment variable EMACS is set to the version of
the running Emacs. I was using it as the path to the emacs
executable. The variable has been renamed in the Makefile to
EMACS_EXE.

PANDOC is defined but not used, as well. I have renamed it for
consistency, but it should be removed.